Add an epi_recipe to a workflow

Add an epi_recipe to a workflow

Usage

add_epi_recipe(x, recipe, ..., blueprint = default_epi_recipe_blueprint())

Arguments

x

A workflow or epi_workflow

recipe

A recipe created using recipes::recipe()

...

Not used.

blueprint

A hardhat blueprint used for fine tuning the preprocessing.

default_epi_recipe_blueprint() is used.

Note that preprocessing done here is separate from preprocessing that might be done automatically by the underlying model.

Value

x, updated with a new recipe preprocessor.

Details

Has the same behaviour as workflows::add_recipe() but sets a different default blueprint to automatically handle epiprocess::epi_df data.

See also

Examples

library(epiprocess)
#> 
#> Attaching package: ‘epiprocess’
#> The following object is masked from ‘package:stats’:
#> 
#>     filter
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
library(recipes)
#> 
#> Attaching package: ‘recipes’
#> The following object is masked from ‘package:stats’:
#> 
#>     step

jhu <- case_death_rate_subset %>%
  filter(time_value > "2021-08-01") %>%
  dplyr::arrange(geo_value, time_value)

r <- epi_recipe(jhu) %>%
  step_epi_lag(death_rate, lag = c(0, 7, 14)) %>%
  step_epi_ahead(death_rate, ahead = 7) %>%
  step_epi_lag(case_rate, lag = c(0, 7, 14)) %>%
  step_naomit(all_predictors()) %>%
  step_naomit(all_outcomes(), skip = TRUE)

workflow <- epi_workflow() %>%
  add_epi_recipe(r)

workflow
#> ══ Epi Workflow ════════════════════════════════════════════════════════════════
#> Preprocessor: Recipe
#> Model: None
#> Postprocessor: None
#> 
#> ── Preprocessor ────────────────────────────────────────────────────────────────
#> 5 Recipe Steps
#> 
#> • step_epi_lag()
#> • step_epi_ahead()
#> • step_epi_lag()
#> • step_naomit()
#> • step_naomit()