ARX forecaster argument constructor

Constructs a list of arguments for arx_forecaster().

Usage

arx_args_list(
  lags = c(0, 7, 14),
  ahead = 7,
  min_train_window = 20,
  levels = c(0.05, 0.95),
  intercept = TRUE,
  symmetrize = TRUE,
  nonneg = TRUE,
  quantile_by_key = FALSE
)

Arguments

lags

Vector or List. Positive integers enumerating lags to use in autoregressive-type models.

ahead

Integer. Number of time steps ahead of the forecast date for which forecasts should be produced.

min_train_window

Integer. The minimal amount of training data needed to produce a forecast. If smaller, the forecaster will return NA predictions.

levels

Vector or NULL. A vector of probabilities to produce prediction intervals. These are created by computing the quantiles of training residuals. A NULL value will result in point forecasts only.

intercept

Logical. The default TRUE includes intercept in the forecaster.

symmetrize

Logical. The default TRUE calculates symmetric prediction intervals.

nonneg

Logical. The default TRUE enforeces nonnegative predictions by hard-thresholding at 0.

quantile_by_key

Not currently implemented

Value

A list containing updated parameter choices.

Examples

arx_args_list()
#> $lags
#> [1]  0  7 14
#> 
#> $ahead
#> [1] 7
#> 
#> $min_train_window
#> [1] 20
#> 
#> $levels
#> [1] 0.05 0.95
#> 
#> $intercept
#> [1] TRUE
#> 
#> $symmetrize
#> [1] TRUE
#> 
#> $nonneg
#> [1] TRUE
#> 
#> $max_lags
#> [1] 14
#> 
arx_args_list(symmetrize = FALSE)
#> $lags
#> [1]  0  7 14
#> 
#> $ahead
#> [1] 7
#> 
#> $min_train_window
#> [1] 20
#> 
#> $levels
#> [1] 0.05 0.95
#> 
#> $intercept
#> [1] TRUE
#> 
#> $symmetrize
#> [1] FALSE
#> 
#> $nonneg
#> [1] TRUE
#> 
#> $max_lags
#> [1] 14
#> 
arx_args_list(levels = c(.1, .3, .7, .9), min_train_window = 120)
#> $lags
#> [1]  0  7 14
#> 
#> $ahead
#> [1] 7
#> 
#> $min_train_window
#> [1] 120
#> 
#> $levels
#> [1] 0.1 0.3 0.7 0.9
#> 
#> $intercept
#> [1] TRUE
#> 
#> $symmetrize
#> [1] TRUE
#> 
#> $nonneg
#> [1] TRUE
#> 
#> $max_lags
#> [1] 14
#>