ARX classifier argument constructor

Constructs a list of arguments for arx_classifier().

Usage

arx_class_args_list(
  lags = c(0L, 7L, 14L),
  ahead = 7L,
  min_train_window = 20L,
  forecast_date = NULL,
  target_date = NULL,
  outcome_transform = c("growth_rate", "lag_difference"),
  breaks = 0.25,
  horizon = 7L,
  method = c("rel_change", "linear_reg", "smooth_spline", "trend_filter"),
  log_scale = FALSE,
  additional_gr_args = list()
)

Arguments

lags

Vector or List. Positive integers enumerating lags to use in autoregressive-type models (in days).

ahead

Integer. Number of time steps ahead (in days) of the forecast date for which forecasts should be produced.

min_train_window

Integer. The minimal amount of training data (in the time unit of the epi_df) needed to produce a forecast. If smaller, the forecaster will return NA predictions.

forecast_date

Date. The date on which the forecast is created. The default NULL will attempt to determine this automatically.

target_date

Date. The date for which the forecast is intended. The default NULL will attempt to determine this automatically.

outcome_transform

Scalar character. Whether the outcome should be created using growth rates (as the predictors are) or lagged differences or growth rates. The second case is closer to the requirements for the 2022-23 CDC Flusight Hospitalization Experimental Target. See the Classification Vignette for details of how to create a reasonable baseline for this case.

breaks

Vector. A vector of breaks to turn real-valued growth rates into discrete classes. The default gives binary upswing classification as in McDonald, Bien, Green, Hu, et al.. This coincides with the default trainer = parsnip::logistic_reg() argument in arx_classifier(). However, multiclass classification is also supported (e.g. with breaks = c(-.2, .25)) provided that trainer = parsnip::multinom_reg() (or another multiclass trainer) is used as well. These will be sliently expanded to cover the entire real line (so the default will become breaks = c(-Inf, .25, Inf)) before being used to discretize the response. This is different than the behaviour in recipes::step_cut() which creates classes that only cover the range of the training data.

horizon

Scalar integer. This is passed to the h argument of epiprocess::growth_rate(). It determines the amount of data used to calculate the growth rate.

method

Character. Options available for growth rate calculation.

log_scale

Scalar logical. Whether to compute growth rates on the log scale.

additional_gr_args

List. Optional arguments controlling growth rate calculation. See epiprocess::growth_rate() and the related Vignette for more details.

Value

A list containing updated parameter choices with class arx_alist.

Examples

arx_class_args_list()
#> $lags
#> [1]  0  7 14
#> 
#> $ahead
#> [1] 7
#> 
#> $min_train_window
#> [1] 20
#> 
#> $breaks
#> [1] -Inf 0.25  Inf
#> 
#> $forecast_date
#> NULL
#> 
#> $target_date
#> NULL
#> 
#> $outcome_transform
#> [1] "growth_rate"
#> 
#> $max_lags
#> [1] 14
#> 
#> $horizon
#> [1] 7
#> 
#> $method
#> [1] "rel_change"
#> 
#> $log_scale
#> [1] FALSE
#> 
#> $additional_gr_args
#> list()
#> 
#> attr(,"class")
#> [1] "arx_clist"

# 3-class classsification,
# also needs arx_classifier(trainer = parsnip::multinom_reg())
arx_class_args_list(breaks = c(-.2, .25))
#> $lags
#> [1]  0  7 14
#> 
#> $ahead
#> [1] 7
#> 
#> $min_train_window
#> [1] 20
#> 
#> $breaks
#> [1]  -Inf -0.20  0.25   Inf
#> 
#> $forecast_date
#> NULL
#> 
#> $target_date
#> NULL
#> 
#> $outcome_transform
#> [1] "growth_rate"
#> 
#> $max_lags
#> [1] 14
#> 
#> $horizon
#> [1] 7
#> 
#> $method
#> [1] "rel_change"
#> 
#> $log_scale
#> [1] FALSE
#> 
#> $additional_gr_args
#> list()
#> 
#> attr(,"class")
#> [1] "arx_clist"