Direct autoregressive forecaster with covariates

This is an autoregressive forecasting model for epiprocess::epi_df data. It does "direct" forecasting, meaning that it estimates a model for a particular target horizon.

Usage

arx_forecaster(
  epi_data,
  outcome,
  predictors,
  trainer = parsnip::linear_reg(),
  args_list = arx_args_list()
)

Arguments

epi_data

An epi_df object

outcome

A character (scalar) specifying the outcome (in the epi_df).

predictors

A character vector giving column(s) of predictor variables.

trainer

A {parsnip} model describing the type of estimation. For now, we enforce mode = "regression".

args_list

A list of customization arguments to determine the type of forecasting model. See arx_args_list().

Value

A list with (1) predictions an epi_df of predicted values and (2) epi_workflow, a list that encapsulates the entire estimation workflow

See also

Examples

jhu <- case_death_rate_subset %>%
  dplyr::filter(time_value >= as.Date("2021-12-01"))

out <- arx_forecaster(jhu, "death_rate",
  c("case_rate", "death_rate"))
#> Warning: The forecast_date is less than the most recent update date of the data.forecast_date = 2021-12-31 while data is from 2022-05-31.

out <- arx_forecaster(jhu, "death_rate",
  c("case_rate", "death_rate"), trainer = quantile_reg(),
  args_list = arx_args_list(levels = 1:9 / 10))
#> Warning: The forecast_date is less than the most recent update date of the data.forecast_date = 2021-12-31 while data is from 2022-05-31.