Produce a forecast from an epi workflow

Produce a forecast from an epi workflow

Usage

# S3 method for class 'epi_workflow'
forecast(object, ..., n_recent = NULL, forecast_date = NULL)

Arguments

object

An epi workflow.

...

Not used.

n_recent

Integer or NULL. If filling missing data with locf = TRUE, how far back are we willing to tolerate missing data? Larger values allow more filling. The default NULL will determine this from the the recipe. For example, suppose n_recent = 3, then if the 3 most recent observations in any geo_value are all NA’s, we won’t be able to fill anything, and an error message will be thrown. (See details.)

forecast_date

By default, this is set to the maximum time_value in x. But if there is data latency such that recent NA's should be filled, this may be after the last available time_value.

Value

A forecast tibble.