Get the names from a data frame via tidy select

Given a data.frame, use <tidy-select> syntax to choose some variables. Return the names of those variables

Usage

grab_names(dat, ...)

Arguments

dat

a data.frame

...

<tidy-select> One or more unquoted expressions separated by commas. Variable names can be used as if they were positions in the data frame, so expressions like x:y can be used to select a range of variables.

Value

a character vector

Details

As this is an internal function, no checks are performed.

Examples

df <- data.frame(a = 1, b = 2, cc = rep(NA, 3))
grab_names(df, dplyr::starts_with("c"))
#> [1] "cc"