Predict from an epi_workflow

This is the predict() method for a fit epi_workflow object. The nice thing about predicting from an epi_workflow is that it will:

Preprocess new_data using the preprocessing method specified when the workflow was created and fit. This is accomplished using hardhat::forge(), which will apply any formula preprocessing or call recipes::bake() if a recipe was supplied.

Call parsnip::predict.model_fit() for you using the underlying fit parsnip model.

Ensure that the returned object is an epiprocess::epi_df where possible. Specifically, the output will have time_value and geo_value columns as well as the prediction.

Usage

# S3 method for epi_workflow
predict(object, new_data, ...)

Arguments

object

An epi_workflow that has been fit by workflows::fit.workflow()

new_data

A data frame containing the new predictors to preprocess and predict on

...

Arguments to the underlying model's prediction function cannot be passed here (see opts). There are some parsnip related options that can be passed, depending on the value of type. Possible arguments are:

interval: for types of "survival" and "quantile", should interval estimates be added, if available? Options are "none" and "confidence".

level: for types of "conf_int", "pred_int", and "survival" this is the parameter for the tail area of the intervals (e.g. confidence level for confidence intervals). Default value is 0.95.

std_error: add the standard error of fit or prediction (on the scale of the linear predictors) for types of "conf_int" and "pred_int". Default value is FALSE.

quantile: the quantile(s) for quantile regression (not implemented yet)

time: the time(s) for hazard and survival probability estimates.

Value

A data frame of model predictions, with as many rows as new_data has. If new_data is an epi_df or a data frame with time_value or geo_value columns, then the result will have those as well.

Examples

library(dplyr)
library(parsnip)
library(recipes)

jhu <- case_death_rate_subset

r <- epi_recipe(jhu) %>%
  step_epi_lag(death_rate, lag = c(0, 7, 14)) %>%
  step_epi_ahead(death_rate, ahead = 7) %>%
  step_epi_lag(case_rate, lag = c(0, 7, 14)) %>%
  step_naomit(all_predictors()) %>%
  step_naomit(all_outcomes(), skip = TRUE)

wf <- epi_workflow(r, linear_reg()) %>% fit(jhu)

latest <- jhu %>% filter(time_value >= max(time_value) - 14)

preds <- predict(wf, latest) %>%
  filter(!is.na(.pred))

preds
#> An `epi_df` object, 56 x 3 with metadata:
#> * geo_type  = state
#> * time_type = day
#> * as_of     = 2022-05-31 19:08:25
#> 
#> # A tibble: 56 × 3
#>    geo_value time_value     .pred
#>  * <chr>     <date>         <dbl>
#>  1 ak        2021-12-31  0.378   
#>  2 al        2021-12-31  0.280   
#>  3 ar        2021-12-31  0.452   
#>  4 as        2021-12-31 -0.000475
#>  5 az        2021-12-31  0.657   
#>  6 ca        2021-12-31  0.293   
#>  7 co        2021-12-31  0.569   
#>  8 ct        2021-12-31  0.641   
#>  9 dc        2021-12-31  1.02    
#> 10 de        2021-12-31  0.733   
#> # … with 46 more rows