State population data

Data set on state populations, from the 2019 US Census.

Usage

state_census

Format

Data frame with 57 rows (including one for the United States as a whole, plus the District of Columbia, Puerto Rico Commonwealth, American Samoa, Guam, the U.S. Virgin Islands, and the Northern Mariana, Islands).

fips

FIPS code

name

Full name of the state or territory

pop

Estimate of the location's resident population in 2019.

abbr

Postal abbreviation for the location

Source

United States Census Bureau, at https://www2.census.gov/programs-surveys/popest/datasets/2010-2019/counties/totals/co-est2019-alldata.pdf, https://www.census.gov/data/tables/time-series/demo/popest/2010s-total-puerto-rico-municipios.html, and https://www.census.gov/data/tables/2010/dec/2010-island-areas.html