Create a leading outcome

step_epi_ahead creates a specification of a recipe step that will add new columns of leading data. Leading data will by default include NA values where the lag was induced. These can be removed with step_naomit(), or you may specify an alternative filler value with the default argument.

step_epi_lag creates a specification of a recipe step that will add new columns of lagged data. Lagged data will by default include NA values where the lag was induced. These can be removed with step_naomit(), or you may specify an alternative filler value with the default argument.

Usage

step_epi_ahead(
  recipe,
  ...,
  role = "outcome",
  trained = FALSE,
  ahead = 1,
  prefix = "ahead_",
  default = NA,
  keys = epi_keys(recipe),
  columns = NULL,
  skip = FALSE,
  id = rand_id("epi_ahead")
)

step_epi_lag(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  lag = 1,
  prefix = "lag_",
  default = NA,
  keys = epi_keys(recipe),
  columns = NULL,
  skip = FALSE,
  id = rand_id("epi_lag")
)

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

...

One or more selector functions to choose variables for this step. See selections() for more details.

role

For model terms created by this step, what analysis role should they be assigned?

trained

A logical to indicate if the quantities for preprocessing have been estimated.

ahead

A vector of positive integers. Each specified column will be lead for each value in the vector.

prefix

A prefix for generated column names, default to "ahead_".

default

Determines what fills empty rows left by leading/lagging (defaults to NA).

keys

A character vector of the keys in an epi_df

columns

A character string of variable names that will be populated (eventually) by the terms argument.

skip

A logical. Should the step be skipped when the recipe is baked by bake()? While all operations are baked when prep() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

id

A character string that is unique to this step to identify it.

lag

A vector of positive integers. Each specified column will be lagged for each value in the vector.

Value

An updated version of recipe with the new step added to the sequence of any existing operations. An updated version of recipe with the new step added to the sequence of any existing operations.

Details

The step assumes that the data are already in the proper sequential order for leading.

The step assumes that the data are already in the proper sequential order for lagging.

Examples

tib <- tibble::tibble(
  x = 1:5, y = 1:5,
  time_value = seq(as.Date("2020-01-01"), by = 1, length.out = 5),
  geo_value = "ca"
  ) %>% epiprocess::as_epi_df()

library(recipes)
epi_recipe(y ~ x, data = tib) %>%
  step_epi_lag(x, lag = 2:3) %>%
  step_epi_ahead(y, ahead = 1) %>%
  prep(tib) %>%
  bake(tib)
#> An `epi_df` object, with metadata:
#> * geo_type  = state
#> * time_type = day
#> * as_of     = 2022-06-08 00:27:56
#> 
#> # A tibble: 9 × 7
#>       x     y time_value geo_value lag_2_x lag_3_x ahead_1_y
#>   <int> <int> <date>     <fct>       <int>   <int>     <int>
#> 1    NA    NA 2019-12-31 ca             NA      NA         1
#> 2     1     1 2020-01-01 ca             NA      NA         2
#> 3     2     2 2020-01-02 ca             NA      NA         3
#> 4     3     3 2020-01-03 ca              1      NA         4
#> 5     4     4 2020-01-04 ca              2       1         5
#> 6     5     5 2020-01-05 ca              3       2        NA
#> 7    NA    NA 2020-01-06 ca              4       3        NA
#> 8    NA    NA 2020-01-07 ca              5       4        NA
#> 9    NA    NA 2020-01-08 ca             NA       5        NA