Calculate a lagged difference

step_lag_difference() creates a specification of a recipe step that will generate one or more new columns of derived data.

Usage

step_lag_difference(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  horizon = 7,
  prefix = "lag_diff_",
  columns = NULL,
  skip = FALSE,
  id = rand_id("lag_diff")
)

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

...

One or more selector functions to choose variables for this step. See recipes::selections() for more details.

role

For model terms created by this step, what analysis role should they be assigned? lag is default a predictor while ahead is an outcome.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

horizon

Scalar or vector. Time period(s) over which to calculate differences.

prefix

A prefix to indicate what type of variable this is

columns

A character string of variable names that will be populated (eventually) by the terms argument.

skip

A logical. Should the step be skipped when the recipe is baked by bake()? While all operations are baked when prep() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

id

A unique identifier for the step

Value

An updated version of recipe with the new step added to the sequence of any existing operations.

See also

Examples

r <- epi_recipe(case_death_rate_subset) %>%
  step_lag_difference(case_rate, death_rate, horizon = c(7, 14))
r
#> 
#> ── Epi Recipe ──────────────────────────────────────────────────────────────────
#> 
#> ── Inputs 
#> Number of variables by role
#> raw:        2
#> geo_value:  1
#> time_value: 1
#> 
#> ── Operations 
#> 1. Calculating lag_difference for: case_rate and death_rate by 7, 14

r %>%
  recipes::prep() %>%
  recipes::bake(case_death_rate_subset)
#> Warning: ! No training data was supplied to `prep()`.
#> ! Unlike a <recipe>, an <epi_recipe> does not
#> ! store the full template data in the object.
#> ! Please supply the training data to the `prep()` function,
#> ! to avoid addtional warning messages.
#> An `epi_df` object, 20,496 x 8 with metadata:
#> * geo_type  = state
#> * time_type = day
#> * as_of     = 2022-05-31 19:08:25.791826
#> 
#> # A tibble: 20,496 × 8
#>    geo_value time_value case_rate death_rate lag_diff_7_case_rate
#>  * <chr>     <date>         <dbl>      <dbl>                <dbl>
#>  1 ak        2020-12-31      35.9      0.158                   NA
#>  2 al        2020-12-31      65.1      0.438                   NA
#>  3 ar        2020-12-31      66.0      1.27                    NA
#>  4 as        2020-12-31       0        0                       NA
#>  5 az        2020-12-31      76.8      1.10                    NA
#>  6 ca        2020-12-31      96.0      0.751                   NA
#>  7 co        2020-12-31      35.8      0.649                   NA
#>  8 ct        2020-12-31      52.1      0.819                   NA
#>  9 dc        2020-12-31      31.0      0.601                   NA
#> 10 de        2020-12-31      65.2      0.807                   NA
#> # ℹ 20,486 more rows
#> # ℹ 3 more variables: lag_diff_14_case_rate <dbl>, lag_diff_7_death_rate <dbl>,
#> #   lag_diff_14_death_rate <dbl>